+++ src/crypto/mishmash/dhe_rsa_aes_gcm_sha384.c
@@ -29,15 +29,15 @@ FILE_LICENCE ( GPL2_OR_LATER_OR_UBDL );
 #include <ipxe/sha512.h>
 #include <ipxe/tls.h>
 
-/** TLS_RSA_WITH_AES_256_GCM_SHA384 cipher suite */
+/** TLS_DHE_RSA_WITH_AES_256_GCM_SHA384 cipher suite */
 struct tls_cipher_suite
-tls_rsa_with_aes_256_gcm_sha384 __tls_cipher_suite ( 12 ) = {
-	.code = htons ( TLS_RSA_WITH_AES_256_GCM_SHA384 ),
+tls_dhe_rsa_with_aes_256_gcm_sha384 __tls_cipher_suite ( 02 ) = {
+	.code = htons ( TLS_DHE_RSA_WITH_AES_256_GCM_SHA384 ),
 	.key_len = ( 256 / 8 ),
 	.fixed_iv_len = 4,
 	.record_iv_len = 8,
 	.mac_len = 0,
-	.exchange = &tls_pubkey_exchange_algorithm,
+	.exchange = &tls_dhe_exchange_algorithm,
 	.pubkey = &rsa_algorithm,
 	.cipher = &aes_gcm_algorithm,
 	.digest = &sha384_algorithm,
